refactor:use a CommonSyncJobTemplate
TIS21-5812
