+++ src/main/java/uk/nhs/tis/sync/job/person/PersonElasticSearchSyncJob.java
@@ -125,7 +125,6 @@ public class PersonElasticSearchSyncJob implements RunnableJob {
     personElasticSearchSync();
   }
 
-
   protected void run() {
 
     if (applicationEventPublisher != null) {
